added messaging for reserving a hall

+++ bioscoop-app-hall-mgmt-service/src/main/java/be/ugent/hallmgmtservice/domain/Seat.java
@@ -42,7 +42,7 @@ public class Seat {
     public void setId(int row, int seat) {
         this.id = Long.parseLong(String.format("%02d%02d", rowNumber, seatNumber));
     }
-    public void setId(int row, int seat, Long eventid) {
+    public void setId(int row, int seat, int eventid) {
         this.id = Long.parseLong(String.format("%02d%02d%02d", rowNumber, seatNumber, eventid));
     }
